Update source URL for srcML binaries
The download location for srcML has changed; oddly
enough, the official site http://www.srcml.org/downloads.html
points to an IP-based (and not name-based) address.
Since the old URL does not provide the required data any
more, switch to the new address.

+++ integration-scripts/install_cppstats.sh
@@ -14,8 +14,16 @@ echo '#!/bin/bash' > $CPPSTATS/cppstats
 echo "cd $CPPSTATS" >> $CPPSTATS/cppstats
 echo "PYTHONPATH=\"\$PYTHONPATH:$CPPSTATS/lib\" ./cppstats.py \"\$@\"" >> $CPPSTATS/cppstats
 chmod +x $CPPSTATS/cppstats
-wget --quiet http://sdml.info/lmcrs/srcML-Ubuntu12.04-64.tar.gz -O /tmp/srcML.tar.gz
-tar -xvf /tmp/srcML.tar.gz
+
+echo "Providing srcML"
+SRCML_URL="http://131.123.42.38/lmcrs/srcML-Ubuntu14.04-64.tar.gz"
+wget --quiet ${SRCML_URL} -O ${TMPDIR}/srcML.tar.gz
+if [ ! -e ${TMPDIR}/srcML.tar.gz ]
+then
+    echo "Could not download srcML from ${SRCML_URL}"
+    exit 1
+fi
+tar -xvf ${TMPDIR}/srcML.tar.gz
 cp -rf $PWD/srcML/* $CPPSTATS/lib/srcml/linux/
 
 sudo ln -sf $CPPSTATS/cppstats /usr/local/bin/cppstats
